Move document panels to inspector folder.
