+++ assets/src/edit-story/components/inspector/document/test/publish.js
@@ -23,9 +23,9 @@ import { ThemeProvider } from 'styled-components';
 /**
  * Internal dependencies
  */
-import StoryContext from '../../../app/story/context';
-import InspectorContext from '../../inspector/context';
-import theme from '../../../theme';
+import StoryContext from '../../../../app/story/context';
+import InspectorContext from '../../../inspector/context';
+import theme from '../../../../theme';
 import PublishPanel from '../publish';
 
 function setupPanel() {
